+++ boards/wdsp-dev/board.mk
@@ -10,9 +10,4 @@ vscode:
 	sed -i'' -e 's#_WDSP_PATH_#$(WDSP_PATH)#;s#_BOARD_#$(BOARD)#;s#_CORE_#$(CORE)#' .vscode/c_cpp_properties.json
 	sed -i'' -e 's#_TARGET_#$(TARGET)#' .vscode/launch.json
 
-.PHONY: skel
-skel:
-	cp $(WDSP_PATH)/boards/$(BOARD)/skel/skel.c ${TARGET}.c
-	cp $(WDSP_PATH)/boards/$(BOARD)/config.ini .
-
 endif
