+++ libwdsp.mk
@@ -14,6 +14,7 @@ CONFIG_FILE ?= $(wildcard ./config.ini)
 export USER_CONFIG ?= $(realpath $(CONFIG_FILE))
 
 include $(WDSP_PATH)/buildvars.mk
+include $(WDSP_PATH)/skel.mk
 
 C_INCLUDES = \
 -I $(WDSP_PATH)/$(CONFIG_DIR) \
@@ -54,3 +55,4 @@ libclean:
 
 .PHONY: clean
 clean: libclean
+
