+++ skel/config.ini
@@ -1,0 +1,33 @@
+; [modules]
+; usb = true
+; midi = true
+
+; [audio]
+; sample_rate = 48000 # Available options are 44100, 45000, 48000 and 50000
+; channels = 2
+; buffer_size = 16
+; block_size = 8
+; usb = true
+
+; [audio_usb]
+; interface_name = "USB Audio Interface"
+; sample_format = s24 # Available options are f32, s24 and s16
+; channels = 2
+; input = true
+; output = true
+; in_pos = pre
+; out_pos = post
+
+; [usb]
+; vid = 0xDEAD
+; pid = 0xC0DE
+; device_version = 0x0100
+; vendor_name = "NLS"
+; product_name = "WhateverDSP Dev Board"
+
+; [midi]
+; uart = true
+; usb = true
+
+; [midi_usb]
+; interface_name = "USB MIDI Interface"
